refactor(models): split schemas.py into per-domain modules
Pure refactor. backend/models/ becomes a package with per-domain
modules (job, dashboard, system, drive, notification, transcoder,
logs, metadata, settings, folder). schemas.py is preserved as a
thin re-export shim so 'from backend.models.schemas import X' keeps
working until callers migrate.

FolderScanRequest and FolderCreateRequest move from routers/folder.py
to models/folder.py for symmetry with the rest of the schema split.

No behavior change. All 643 existing tests pass.

Spec: docs/superpowers/specs/2026-05-03-frontend-type-codegen-design.md
